Refactor: split thermite piece data from piece visuals
thermite/utils.ts co-located the squarePieces data the engine needs
with SVG asset imports, so the engine module couldn't be imported
outside the Next.js build (needed for server-side replay). Moved the
pure piece data + types into pieces.ts (no asset imports); utils.ts
keeps the SVGs as a pieceImages map keyed by piece type. Also drops
dead helpers (getRandomPiece, getRandomSquare, initialBoard,
getStatusMessage).
